+++ app/lib/bk/compat/models/pipeline.rb
@@ -4,7 +4,6 @@ require_relative '../renderer'
 
 # for easier usage (modules just have to import)
 require_relative 'pipeline'
-require_relative '../pipeline/step'
 
 module BK
   module Compat
