+++ app/lib/bk/compat/models/steps/wait.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'base'
+
 module BK
   module Compat
     # simple waiting step
