+++ app/lib/bk/compat/parsers/bitbucket.rb
@@ -2,7 +2,7 @@
 
 require_relative '../translator'
 require_relative '../models/pipeline'
-require_relative '../pipeline/step'
+require_relative '../models/steps/command'
 
 require_relative 'bitbucket/import'
 require_relative 'bitbucket/parallel'
