+++ app/lib/bk/compat/parsers/bitbucket/services.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require_relative '../../pipeline/step'
-require_relative '../../models/plugin'
-
 module BK
   module Compat
     module BitBucketSteps
