+++ app/lib/bk/compat/parsers/harness.rb
@@ -2,7 +2,7 @@
 
 require_relative '../translator'
 require_relative '../models/pipeline'
-require_relative '../pipeline/step'
+require_relative '../models/steps/group'
 require_relative 'harness/steps'
 
 module BK
