+++ app/lib/bk/compat/parsers/harness/steps/run.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative '../../../pipeline/step'
-
 module BK
   module Compat
     module HarnessSteps
